requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
mpmath>=1.2.0
ipywidgets>=8.0.0
mpl_toolkits.mplot3d